# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schedule",
        "description": "Documentation de l'API Schedule",
        "contact": {
            "email": "wwwmbassiloic@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API "
        }
    ],
    "paths": {
        "/access/": {
            "get": {
                "tags": [
                    "Access"
                ],
                "summary": "Get list of access",
                "description": "Returns list of access",
                "operationId": "getaccessList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/access/{id}": {
            "get": {
                "tags": [
                    "Access"
                ],
                "summary": "Get access information",
                "description": "Returns access data",
                "operationId": "9a7db369fddca170281321a6f7d19499",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "access id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/access/delete/{id}": {
            "delete": {
                "tags": [
                    "Access"
                ],
                "summary": "Delete existing access",
                "description": "Deletes a record and returns no content",
                "operationId": "f24c84619d66e780ce2b95d0c01299ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "access id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/access/save": {
            "post": {
                "tags": [
                    "Access"
                ],
                "summary": "Store new access",
                "description": "Returns access data",
                "operationId": "c4a4422cd198a3d183664f3e785a737d",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/access/update/{id}": {
            "put": {
                "tags": [
                    "Access"
                ],
                "summary": "Update existing access",
                "description": "Returns updated access data",
                "operationId": "bbd0c312cc40eced0d318f342c8eb282",
                "parameters": [
                    {
                        "name": "sccessName",
                        "in": "path",
                        "description": "access name: you should give the access name to update it",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/update/{id}": {
            "put": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Store new project",
                "description": "Returns project data",
                "operationId": "26ed056009a8448597ef9698905dfd59",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/delete-use/{id}": {
            "delete": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
}